+++ python/cxx/group.cpp
@@ -1,23 +1,23 @@
-#include <opm/parser/eclipse/EclipseState/Schedule/Group/Group.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/Group/Group2.hpp>
 #include <pybind11/stl.h>
 #include "sunbeam.hpp"
 
 
 namespace {
 
-  const std::set<std::string> wellnames( const Group& g, size_t timestep ) {
-        return g.getWells( timestep );
+  const std::vector<std::string> wellnames( const Group2& g ) {
+        return g.wells( );
     }
 
-     int get_vfp_table_nr( const Group& g, size_t timestep ) {
-        return g.getGroupNetVFPTable(timestep);
+     int get_vfp_table_nr( const Group2& g ) {
+        return g.getGroupNetVFPTable();
     }
 }
 
 void sunbeam::export_Group(py::module& module) {
 
-  py::class_< Group >( module, "Group")
-    .def_property_readonly( "name", &Group::name)
+  py::class_< Group2 >( module, "Group")
+    .def_property_readonly( "name", &Group2::name)
     .def( "_vfp_table_nr", &get_vfp_table_nr )
     .def( "_wellnames", &wellnames );
